Change login scene to a choose scene between mojang and microsoft login (only mojang implemented currently)
