+++ src/main/java/com/troblecodings/launcher/assets/style.css
@@ -251,6 +251,18 @@ CheckBox {
 	-fx-pref-width: 164px;
 }
 
+.mojangLoginButton{
+	-fx-background-image: url(mojang-login-button.png);
+	-fx-pref-height: 42px;
+	-fx-pref-width: 210px;
+}
+
+.microsoftLoginButton{
+	-fx-background-image: url(microsoft-login-button.png);
+	-fx-pref-height: 42px;
+	-fx-pref-width: 210px;
+}
+
 .closebutton {
 	-fx-background-image: url(closebutton.png);
 	-fx-pref-height: 67px;
